+++ Sources/Base/OAuth2CodeGrantNoTokenType.swift
@@ -1,9 +1,9 @@
 //
-//  OAuth2.h
+//  OAuth2CodeGrantNoTokenType.swift
 //  OAuth2
 //
-//  Created by Pascal Pfiffner on 6/4/14.
-//  Copyright 2014 Pascal Pfiffner
+//  Created by Pascal Pfiffner on 1/15/16.
+//  Copyright 2016 Pascal Pfiffner.
 //
 //  Licensed under the Apache License, Version 2.0 (the "License");
 //  you may not use this file except in compliance with the License.
@@ -18,14 +18,16 @@
 //  limitations under the License.
 //
 
-@import Foundation;
 
-//! Project version number for OAuth2.
-FOUNDATION_EXPORT double OAuth2VersionNumber;
-
-//! Project version string for OAuth2.
-FOUNDATION_EXPORT const unsigned char OAuth2VersionString[];
-
-// In this header, you should import all the public headers of your framework using statements like #import <OAuth2/PublicHeader.h>
-
-
+/**
+Subclass to deal with sites that don't return `token_type`, such as Instagram or Bitly.
+*/
+public class OAuth2CodeGrantNoTokenType: OAuth2CodeGrant {
+	
+	public override init(settings: OAuth2JSON) {
+		super.init(settings: settings)
+	}
+	
+	override func assureCorrectBearerType(params: OAuth2JSON) throws {
+	}
+}